Move ClubEventSerializer
